refactor: move getIndicators to a composable
This way we don't need to pass all the data to the method that can easily be retrieves when using other composables.
