+++ packages/web-pkg/src/components/FilesList/ResourceStatusIndicators.vue
@@ -9,11 +9,9 @@
 
 <script setup lang="ts">
 import { Resource, SpaceResource } from '@opencloud-eu/web-client'
-import { getIndicators, ResourceIndicator } from '../../helpers'
 import { computed, useAttrs } from 'vue'
-import { useResourcesStore, useUserStore } from '../../composables/piniaStores'
 import { OcStatusIndicators } from '@opencloud-eu/design-system/components'
-import { useInterceptModifierClick } from '../../composables/keyboardActions'
+import { ResourceIndicator, useResourceIndicators } from '../../composables'
 
 const attrs = useAttrs() as (typeof OcStatusIndicators)['props']
 const {
@@ -26,19 +24,10 @@ const {
   filter?: (indicator: ResourceIndicator) => boolean
 }>()
 
-const userStore = useUserStore()
-const resourcesStore = useResourcesStore()
-const { interceptModifierClick } = useInterceptModifierClick()
+const { getIndicators } = useResourceIndicators()
 
 const indicators = computed(() => {
-  const list = getIndicators({
-    space,
-    resource,
-    ancestorMetaData: resourcesStore.ancestorMetaData,
-    user: userStore.user,
-    interceptModifierClick
-  })
-
+  const list = getIndicators({ space, resource })
   if (filter) {
     return list.filter(filter)
   }
